+++ src/kimi_cli/soul/__init__.py
@@ -16,6 +16,7 @@ from kimi_cli.wire.message import WireMessage
 
 if TYPE_CHECKING:
     from kimi_cli.llm import LLM, ModelCapability
+    from kimi_cli.utils.slashcmd import SlashCommand
 
 
 class LLMNotSet(Exception):
@@ -77,12 +78,18 @@ class Soul(Protocol):
         """The current status of the soul. The returned value is immutable."""
         ...
 
+    @property
+    def available_slash_commands(self) -> list[SlashCommand[Any]]:
+        """List of available slash commands supported by the soul."""
+        ...
+
     async def run(self, user_input: str | list[ContentPart]):
         """
         Run the agent with the given user input until the max steps or no more tool calls.
 
         Args:
             user_input (str | list[ContentPart]): The user input to the agent.
+                Can be a slash command call or natural language input.
 
         Raises:
             LLMNotSet: When the LLM is not set.
@@ -107,7 +114,7 @@ async def run_soul(
     user_input: str | list[ContentPart],
     ui_loop_fn: UILoopFn,
     cancel_event: asyncio.Event,
-    wire_file_backend: Path | None = None,
+    wire_file: Path | None = None,
 ) -> None:
     """
     Run the soul with the given user input, connecting it to the UI loop with a `Wire`.
@@ -122,7 +129,7 @@ async def run_soul(
         MaxStepsReached: When the maximum number of steps is reached.
         RunCancelled: When the run is cancelled by the cancel event.
     """
-    wire = Wire(file_backend=wire_file_backend)
+    wire = Wire(file_backend=wire_file)
     wire_token = _current_wire.set(wire)
 
     logger.debug("Starting UI loop with function: {ui_loop_fn}", ui_loop_fn=ui_loop_fn)